put main minian fucntion in separet file
